Allow fallback max resolution for VA to be read from file
This patch allows the max resolution for VAAPI to be read from a file `/etc/chromium-browser/va_fallback_resolution` in case VAAPI cannot return the actual value on its own, a problem which manifests in Nvidia GPUs using VDPAU.

The file is expected to have two space separated integers in the first line, width and height in that order. In case the file does not exist, or can't be read for whatever reason, default values of 1920x1088 are used.